Revert "[td] Data product orchestration (#3155)"
This reverts commit 608067cba8a6f5162572ef8de9156f32eae1d898.
