+++ mage_ai/api/presenters/PipelineSchedulePresenter.py
@@ -5,7 +5,6 @@ from mage_ai.api.presenters.BasePresenter import BasePresenter
 class PipelineSchedulePresenter(BasePresenter):
     default_attributes = [
         'created_at',
-        'global_data_product_uuid',
         'id',
         'name',
         'pipeline_uuid',
