+++ mage_ai/api/resources/PipelineScheduleResource.py
@@ -28,10 +28,7 @@ class PipelineScheduleResource(DatabaseResource):
                 query.
                 options(selectinload(PipelineSchedule.event_matchers)).
                 options(selectinload(PipelineSchedule.pipeline_runs)).
-                filter(
-                    PipelineSchedule.global_data_product_uuid is None,
-                    PipelineSchedule.pipeline_uuid == pipeline.uuid,
-                ).
+                filter(PipelineSchedule.pipeline_uuid == pipeline.uuid).
                 order_by(PipelineSchedule.id.desc(), PipelineSchedule.start_time.desc())
             )
