--- mage_ai/data_preparation/models/global_data_product/constants.py
@@ -1,6 +1,0 @@
-import enum
-
-
-class GlobalDataProductObjectType(str, enum.Enum):
-    BLOCK = 'block'
-    PIPELINE = 'pipeline'
