+++ mage_ai/frontend/pages/pipelines/[pipeline]/runs/index.tsx
@@ -169,12 +169,7 @@ function PipelineRuns({
     mutate: fetchPipelineRuns,
   } = api.pipeline_runs.list(
     pipelineRunsRequestQuery,
-    {
-      refreshInterval: 5000,
-    },
-    {
-      pauseFetch: !pipelineUUID,
-    },
+    { refreshInterval: 5000 },
   );
   const pipelineRuns: PipelineRunType[] = useMemo(() => dataPipelineRuns?.pipeline_runs || [], [dataPipelineRuns]);
   const totalRuns: number = useMemo(() => isPipelineRunsTab
