+++ mage_ai/orchestration/triggers/constants.py
@@ -1,3 +1,2 @@
 DEFAULT_POLL_INTERVAL = 60
-TRIGGER_NAME_FOR_GLOBAL_DATA_PRODUCT = 'GLOBAL_DATA_PRODUCT_TRIGGER'
 TRIGGER_NAME_FOR_TRIGGER_CREATED_FROM_CODE = 'API_TRIGGER_CREATED_FROM_CODE'
